requirements.txt
ib_async
asyncio
sqlalchemy
fastapi
feedparser
python-dotenv
pytz
pathlib
requests
datetime
werkzeug
dataclasses
setuptools
waitress
uvicorn 
pydantic 
jinja2 
python-multipart
aiofiles